Clarify dev setup
The README isn't totally clear about what is running on your local
machine vs what is running on a docker instance. Adding some more
documentation that will hopefully make that a bit more clear. As a
little side edition this also pins the version of webpack being used to
the ancient one that was initially used when the client was being built.